refactor: split dartboard into components
