+++ src/components/dartboard/StunningDartboard.js
@@ -1,5 +1,8 @@
 'use client'
 import React, { useState, useEffect, Fragment } from 'react'
+import PlayerManagement from './PlayerManagement';
+import ScoreboardUI from './ScoreboardUI';
+import DartboardSVG from './DartboardSVG';
 
 // Main component for the Stunning Dartboard application
 const StunningDartboard = () => {
@@ -2061,1029 +2064,58 @@ const StunningDartboard = () => {
           )}
         </div>
       )}
-      {currentTab === 'players' && (
-        <div className="player-management">
-          <div className="cricket-header">
-            <h2 className="cricket-title">PLAYER MANAGEMENT</h2>
-          </div>
-
-          <div className="add-player-section">
-            <input
-              type="text"
-              className="player-input"
-              placeholder="Enter player name..."
-              value={newPlayerName}
-              onChange={(e) => setNewPlayerName(e.target.value)}
-              onKeyPress={(e) => e.key === 'Enter' && addPlayer()}
-            />
-            <button className="add-btn" onClick={addPlayer}>
-              ADD PLAYER
-            </button>
-          </div>
-
-          <div className="player-selection">
-            <div className="selection-title">SELECT PLAYERS FOR GAME</div>
-            <div className="selection-grid">
-              <div className="player-slot">
-                <div className="slot-title">PLAYER 1</div>
-                <div className="selected-player">
-                  {selectedPlayers.player1
-                    ? selectedPlayers.player1.name
-                    : 'No player selected'}
-                </div>
-                <button
-                  className="control-btn"
-                  onClick={() => selectPlayer('player1', null)}
-                >
-                  CLEAR
-                </button>
-              </div>
-              <div className="player-slot">
-                <div className="slot-title">PLAYER 2</div>
-                <div className="selected-player">
-                  {selectedPlayers.player2
-                    ? selectedPlayers.player2.name
-                    : 'No player selected'}
-                </div>
-                <button
-                  className="control-btn"
-                  onClick={() => selectPlayer('player2', null)}
-                >
-                  CLEAR
-                </button>
-              </div>
-            </div>
-            <div className="bot-section">
-              <h3 className="selection-title">PLAY VS. AI OPPONENT</h3>
-              <div className="bot-slider">
-                <label htmlFor="bot-difficulty">
-                  Bot Difficulty (Avg. Score:{' '}
-                  {botOpponent ? botOpponent.difficulty : 'N/A'})
-                </label>
-                <input
-                  type="range"
-                  id="bot-difficulty"
-                  min="30"
-                  max="90"
-                  step="5"
-                  defaultValue="60"
-                  onChange={(e) =>
-                    setBotOpponent({ difficulty: parseInt(e.target.value) })
-                  }
-                />
-                <button
-                  className="control-btn"
-                  onClick={() => {
-                    const botPlayer = {
-                      id: 'bot',
-                      name: `BOT (${botOpponent?.difficulty || 60})`,
-                      isBot: true,
-                      stats: {
-                        gamesPlayed: 0,
-                        gamesWon: 0,
-                        cricketGames: 0,
-                        cricketWins: 0,
-                        countdownGames: 0,
-                        countdownWins: 0,
-                        totalDarts: 0,
-                        totalScore: 0,
-                        bullsHit: 0,
-                        doublesHit: 0,
-                        triplesHit: 0,
-                        highestFinish: 0,
-                        currentStreak: 0,
-                        longestStreak: 0,
-                        threeDartAverage: 0,
-                        first9Avg: 0,
-                        highestScore: 0,
-                        total180s: 0,
-                        marksPerRound: 0,
-                        checkoutAttempts: 0,
-                        checkoutHits: 0,
-                        checkoutPercentage: 0,
-                      },
-                    }
-                    selectPlayer('player2', botPlayer)
-                  }}
-                >
-                  Start Bot Match
-                </button>
-              </div>
-            </div>
-          </div>
-
-          {selectedPlayers.player1 && selectedPlayers.player2 ? (
-            <div className="player-comparison">
-              <div className="comparison-header">
-                <div className="comparison-title">PLAYER COMPARISON</div>
-                <div className="comparison-subtitle">
-                  {selectedPlayers.player1.name} vs{' '}
-                  {selectedPlayers.player2.name}
-                </div>
-              </div>
-              <div className="charts-grid">
-                <div className="chart-container">
-                  <div className="chart-title">Checkout %</div>
-                  <div className="bar-chart">
-                    {(() => {
-                      const data = generateBarChartData(
-                        'checkoutPercentage',
-                        selectedPlayers.player1,
-                        selectedPlayers.player2,
-                      )
-                      return (
-                        <>
-                          <div className="bar-group">
-                            <div
-                              className="bar"
-                              style={{
-                                '--color': '#22c55e',
-                                height: `${data.player1.height}px`,
-                              }}
-                            >
-                              <div className="bar-value">
-                                {data.player1.value.toFixed(1)}%
-                              </div>
-                            </div>
-                            <div
-                              className="bar-label"
-                              style={{ '--color': '#22c55e' }}
-                            >
-                              {selectedPlayers.player1.name}
-                            </div>
-                          </div>
-                          <div className="bar-group">
-                            <div
-                              className="bar"
-                              style={{
-                                '--color': '#06b6d4',
-                                height: `${data.player2.height}px`,
-                              }}
-                            >
-                              <div className="bar-value">
-                                {data.player2.value.toFixed(1)}%
-                              </div>
-                            </div>
-                            <div
-                              className="bar-label"
-                              style={{ '--color': '#06b6d4' }}
-                            >
-                              {selectedPlayers.player2.name}
-                            </div>
-                          </div>
-                        </>
-                      )
-                    })()}
-                  </div>
-                </div>
-
-                <div className="chart-container">
-                  <div className="chart-title">3-DART AVERAGE</div>
-                  <div className="bar-chart">
-                    {(() => {
-                      const data = generateBarChartData(
-                        'threeDartAverage',
-                        selectedPlayers.player1,
-                        selectedPlayers.player2,
-                      )
-                      return (
-                        <>
-                          <div className="bar-group">
-                            <div
-                              className="bar"
-                              style={{
-                                '--color': '#8b5cf6',
-                                height: `${data.player1.height}px`,
-                              }}
-                            >
-                              <div className="bar-value">
-                                {data.player1.value.toFixed(1)}
-                              </div>
-                            </div>
-                            <div
-                              className="bar-label"
-                              style={{ '--color': '#8b5cf6' }}
-                            >
-                              {selectedPlayers.player1.name}
-                            </div>
-                          </div>
-                          <div className="bar-group">
-                            <div
-                              className="bar"
-                              style={{
-                                '--color': '#ef4444',
-                                height: `${data.player2.height}px`,
-                              }}
-                            >
-                              <div className="bar-value">
-                                {data.player2.value.toFixed(1)}
-                              </div>
-                            </div>
-                            <div
-                              className="bar-label"
-                              style={{ '--color': '#ef4444' }}
-                            >
-                              {selectedPlayers.player2.name}
-                            </div>
-                          </div>
-                        </>
-                      )
-                    })()}
-                  </div>
-                </div>
-
-                <div className="chart-container radar-chart-container">
-                  <div className="chart-title">OVERALL PERFORMANCE</div>
-                  <div className="radar-chart">
-                    <svg className="radar-svg" viewBox="0 0 250 250">
-                      <path
-                        d={generateRadarGridPath()}
-                        className="radar-grid"
-                      />
-                      <path d={generateRadarAxes()} className="radar-axis" />
-                      {generateRadarData(selectedPlayers.player1).map(
-                        (point, index) => (
-                          <text
-                            key={index}
-                            x={point.labelX}
-                            y={point.labelY}
-                            className="radar-label"
-                            dy="0.3em"
-                          >
-                            {point.label}
-                          </text>
-                        ),
-                      )}
-                      <polygon
-                        points={generateRadarData(selectedPlayers.player1)
-                          .map((p) => `${p.x},${p.y}`)
-                          .join(' ')}
-                        className="radar-area"
-                        style={{ '--color': '#22c55e' }}
-                      />
-                      <polygon
-                        points={generateRadarData(selectedPlayers.player2)
-                          .map((p) => `${p.x},${p.y}`)
-                          .join(' ')}
-                        className="radar-area"
-                        style={{ '--color': '#06b6d4' }}
-                      />
-                      {generateRadarData(selectedPlayers.player1).map(
-                        (point, index) => (
-                          <circle
-                            key={`p1-${index}`}
-                            cx={point.x}
-                            cy={point.y}
-                            className="radar-point"
-                            style={{ '--color': '#22c55e' }}
-                          />
-                        ),
-                      )}
-                      {generateRadarData(selectedPlayers.player2).map(
-                        (point, index) => (
-                          <circle
-                            key={`p2-${index}`}
-                            cx={point.x}
-                            cy={point.y}
-                            className="radar-point"
-                            style={{ '--color': '#06b6d4' }}
-                          />
-                        ),
-                      )}
-                    </svg>
-                  </div>
-                  <div className="legend">
-                    <div className="legend-item">
-                      <div
-                        className="legend-color"
-                        style={{ '--color': '#22c55e' }}
-                      ></div>
-                      <div className="legend-text">
-                        {selectedPlayers.player1.name}
-                      </div>
-                    </div>
-                    <div className="legend-item">
-                      <div
-                        className="legend-color"
-                        style={{ '--color': '#06b6d4' }}
-                      ></div>
-                      <div className="legend-text">
-                        {selectedPlayers.player2.name}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          ) : (
-            <div className="player-comparison">
-              <div className="no-comparison">
-                <div style={{ fontSize: '48px', marginBottom: '1rem' }}>📊</div>
-                <div style={{ fontSize: '18px', marginBottom: '0.5rem' }}>
-                  Select two players to see comparison charts
-                </div>
-                <div style={{ fontSize: '14px' }}>
-                  Choose players from the cards below to visualize their
-                  performance
-                </div>
-              </div>
-            </div>
-          )}
-
-          <div className="players-grid">
-            {players.map((player) => (
-              <div
-                key={player.id}
-                className={`player-card ${selectedPlayers.player1?.id === player.id || selectedPlayers.player2?.id === player.id ? 'selected' : ''}`}
-              >
-                <div className="player-header">
-                  {editingPlayer === player.id ? (
-                    <input
-                      type="text"
-                      className="player-input"
-                      defaultValue={player.name}
-                      onKeyPress={(e) =>
-                        e.key === 'Enter' &&
-                        updatePlayerName(player.id, e.target.value)
-                      }
-                      onBlur={(e) =>
-                        updatePlayerName(player.id, e.target.value)
-                      }
-                      autoFocus
-                    />
-                  ) : (
-                    <div className="player-name">{player.name}</div>
-                  )}
-                  <div className="player-actions">
-                    <button
-                      className="action-btn edit-btn"
-                      onClick={() => setEditingPlayer(player.id)}
-                    >
-                      EDIT
-                    </button>
-                    <button
-                      className="action-btn delete-btn"
-                      onClick={() => deletePlayer(player.id)}
-                    >
-                      DEL
-                    </button>
-                    <button
-                      className="action-btn select-btn"
-                      onClick={() => {
-                        if (
-                          selectedPlayers.player1?.id === player.id ||
-                          selectedPlayers.player2?.id === player.id
-                        )
-                          return
-                        if (!selectedPlayers.player1)
-                          selectPlayer('player1', player)
-                        else if (!selectedPlayers.player2)
-                          selectPlayer('player2', player)
-                        else selectPlayer('player1', player) // Replace player 1 if both are full
-                      }}
-                      disabled={
-                        selectedPlayers.player1?.id === player.id ||
-                        selectedPlayers.player2?.id === player.id
-                      }
-                    >
-                      {selectedPlayers.player1?.id === player.id ||
-                      selectedPlayers.player2?.id === player.id
-                        ? 'SELECTED'
-                        : 'SELECT'}
-                    </button>
-                  </div>
-                </div>
-                <div className="player-stats">
-                  <div className="stat-item">
-                    <span className="stat-label">Win %</span>
-                    <span className="stat-value">
-                      {player.stats.gamesPlayed > 0
-                        ? Math.round(
-                            (player.stats.gamesWon / player.stats.gamesPlayed) *
-                              100,
-                          )
-                        : 0}
-                      %
-                    </span>
-                  </div>
-                  <div className="stat-item">
-                    <span className="stat-label">3-Dart Avg</span>
-                    <span className="stat-value">
-                      {player.stats.threeDartAverage.toFixed(1)}
-                    </span>
-                  </div>
-                  <div className="stat-item">
-                    <span className="stat-label">Checkout %</span>
-                    <span className="stat-value">
-                      {player.stats.checkoutPercentage.toFixed(1)}%
-                    </span>
-                  </div>
-                  <div className="stat-item">
-                    <span className="stat-label">Highest Score</span>
-                    <span className="stat-value">
-                      {player.stats.highestScore}
-                    </span>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
+      <PlayerManagement
+        newPlayerName={newPlayerName}
+        setNewPlayerName={setNewPlayerName}
+        addPlayer={addPlayer}
+        players={players}
+        selectedPlayers={selectedPlayers}
+        selectPlayer={selectPlayer}
+        botOpponent={botOpponent}
+        setBotOpponent={setBotOpponent}
+        editingPlayer={editingPlayer}
+        setEditingPlayer={setEditingPlayer}
+        deletePlayer={deletePlayer}
+        updatePlayerName={updatePlayerName}
+        generateBarChartData={generateBarChartData}
+        generateRadarData={generateRadarData}
+        generateRadarGridPath={generateRadarGridPath}
+        generateRadarAxes={generateRadarAxes}
+      />
       {currentTab === 'game' && (
-        <>
-          <div className="cricket-scoreboard">
-            <div className="cricket-header">
-              <h2 className="game-selection-title">GAME SELECTION</h2>
-            </div>
-            {(!selectedPlayers.player1 ||
-              (partyGames.find((g) => g.id === gameMode) &&
-                !selectedPlayers.player2)) && (
-              <div
-                style={{
-                  textAlign: 'center',
-                  padding: '1rem',
-                  background: 'rgba(245,158,11,0.1)',
-                  border: '1px solid rgba(245,158,11,0.3)',
-                  borderRadius: '12px',
-                  marginBottom: '1rem',
-                  color: '#fcd34d',
-                  fontFamily: 'Orbitron, monospace',
-                  fontSize: '14px',
-                }}
-              >
-                ⚠️ Go to PLAYERS tab to select{' '}
-                {partyGames.find((g) => g.id === gameMode) ||
-                standardGames.find((g) => g.id === gameMode)
-                  ? 'two players'
-                  : 'a player'}
-                .
-              </div>
-            )}
-            <div>
-              <div className="game-section-headers-grid">
-                <div
-                  className="game-section-header"
-                  onClick={() => toggleSection('standard')}
-                >
-                  <GameSelectionIcon
-                    type="standard"
-                    expanded={activeGameSection === 'standard'}
-                  />
-                </div>
-                <div
-                  className="game-section-header"
-                  onClick={() => toggleSection('party')}
-                >
-                  <GameSelectionIcon
-                    type="party"
-                    expanded={activeGameSection === 'party'}
-                  />
-                </div>
-                <div
-                  className="game-section-header"
-                  onClick={() => toggleSection('practice')}
-                >
-                  <GameSelectionIcon
-                    type="practice"
-                    expanded={activeGameSection === 'practice'}
-                  />
-                </div>
-              </div>
-
-              <div
-                className={`game-mode-selector ${activeGameSection === 'standard' ? 'expanded' : ''}`}
-              >
-                {standardGames.map((game) => (
-                  <button
-                    key={game.id}
-                    className={`mode-btn ${gameMode === game.id ? 'active' : ''}`}
-                    onClick={() => setGameModeAndReset(game.id)}
-                  >
-                    {game.name}
-                  </button>
-                ))}
-              </div>
-              <div
-                className={`game-mode-selector ${activeGameSection === 'party' ? 'expanded' : ''}`}
-              >
-                {partyGames.map((game) => (
-                  <button
-                    key={game.id}
-                    className={`mode-btn ${gameMode === game.id ? 'active' : ''}`}
-                    onClick={() => setGameModeAndReset(game.id)}
-                  >
-                    {game.name}
-                  </button>
-                ))}
-              </div>
-              <div
-                className={`game-mode-selector ${activeGameSection === 'practice' ? 'expanded' : ''}`}
-              >
-                {practiceModes.map((game) => (
-                  <button
-                    key={game.id}
-                    className={`mode-btn ${gameMode === game.id ? 'active' : ''}`}
-                    onClick={() => setGameModeAndReset(game.id)}
-                  >
-                    {game.name}
-                  </button>
-                ))}
-                <button
-                  className={`mode-btn ${gameMode === 'freeplay' ? 'active' : ''}`}
-                  onClick={() => setGameModeAndReset('freeplay')}
-                >
-                  Free Play
-                </button>
-              </div>
-            </div>
-          </div>
-
-          {isGameWon() && (
-            <div className="winner-display">
-              <div className="winner-text">
-                🏆{' '}
-                {isGameWon() === 'Draw'
-                  ? 'GAME DRAWN!'
-                  : isGameWon() === 'Practice Over'
-                    ? 'PRACTICE COMPLETE'
-                    : `${gameState[isGameWon().toLowerCase().replace(' ', '')].name} WINS!`}{' '}
-                🏆
-              </div>
-            </div>
-          )}
-
-          {gameMode === 'killer' && <KillerScoreboard />}
-
-          {gameMode === 'cricket' && (
-            <div className="cricket-scoreboard">
-              <div className="cricket-header">
-                <h2 className="cricket-title">CRICKET SCOREBOARD</h2>
-                <div className="game-controls">
-                  <button className="control-btn" onClick={resetGame}>
-                    RESET
-                  </button>
-                  <button className="control-btn" onClick={() => nextPlayer()}>
-                    NEXT PLAYER
-                  </button>
-                  <button
-                    className="control-btn undo-btn"
-                    onClick={undoLastDart}
-                    disabled={gameHistory.length === 0}
-                  >
-                    UNDO
-                  </button>
-                  <button
-                    className="control-btn back-btn"
-                    onClick={goBackTurn}
-                    disabled={gameHistory.length === 0}
-                  >
-                    BACK TURN
-                  </button>
-                </div>
-              </div>
-              <div className="last-dart-display">
-                {lastDart ? (
-                  <div>
-                    <span className="last-dart-text">LAST DART:</span>
-                    <span className="last-dart-value">{lastDart}</span>
-                  </div>
-                ) : (
-                  <span className="last-dart-text">
-                    Waiting for dart throw...
-                  </span>
-                )}
-              </div>
-              <div className="cricket-board">
-                <div className="cricket-number">NUM</div>
-                <div
-                  className={`player-column ${currentPlayer === 1 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player1.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 1 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 1 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="player-score">{gameState.player1.score}</div>
-                </div>
-                <div
-                  className={`player-column ${currentPlayer === 2 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player2.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 2 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 2 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="player-score">{gameState.player2.score}</div>
-                </div>
-                {cricketNumbers.concat(['bull']).map((num) => (
-                  <Fragment key={num}>
-                    <div className="cricket-number">
-                      {num === 'bull' ? 'BULL' : num}
-                    </div>
-                    <div
-                      className={`marks-display ${gameState.player1.marks[num] === 3 ? 'closed' : ''}`}
-                    >
-                      {getMarkDisplay(gameState.player1.marks[num])}
-                    </div>
-                    <div
-                      className={`marks-display ${gameState.player2.marks[num] === 3 ? 'closed' : ''}`}
-                    >
-                      {getMarkDisplay(gameState.player2.marks[num])}
-                    </div>
-                  </Fragment>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {['301', '501', '701'].includes(gameMode) && (
-            <div className="cricket-scoreboard">
-              <div className="cricket-header">
-                <h2 className="cricket-title">{gameMode} SCOREBOARD</h2>
-                <div className="game-controls">
-                  <button className="control-btn" onClick={resetGame}>
-                    RESET
-                  </button>
-                  <button className="control-btn" onClick={() => nextPlayer()}>
-                    NEXT PLAYER
-                  </button>
-                  <button
-                    className="control-btn undo-btn"
-                    onClick={undoLastDart}
-                    disabled={gameHistory.length === 0}
-                  >
-                    UNDO
-                  </button>
-                  <button
-                    className="control-btn back-btn"
-                    onClick={goBackTurn}
-                    disabled={gameHistory.length === 0}
-                  >
-                    BACK TURN
-                  </button>
-                </div>
-              </div>
-              <div className="last-dart-display">
-                {lastDart ? (
-                  <div>
-                    <span className="last-dart-text">LAST DART:</span>
-                    <span className="last-dart-value">{lastDart}</span>
-                  </div>
-                ) : (
-                  <span className="last-dart-text">
-                    Waiting for dart throw...
-                  </span>
-                )}
-              </div>
-              <div className="countdown-scoreboard">
-                <div
-                  className={`countdown-player ${currentPlayer === 1 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player1.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 1 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 1 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="countdown-score">
-                    {gameState.player1.score}
-                  </div>
-                  <div className="checkout-suggestion">
-                    {currentPlayer === 1 &&
-                      checkoutSuggestion.length > 0 &&
-                      checkoutSuggestion.map((s, i) => (
-                        <span key={i}>{s}</span>
-                      ))}
-                  </div>
-                </div>
-                <div
-                  className={`countdown-player ${currentPlayer === 2 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player2.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 2 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 2 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="countdown-score">
-                    {gameState.player2.score}
-                  </div>
-                  <div className="checkout-suggestion">
-                    {currentPlayer === 2 &&
-                      checkoutSuggestion.length > 0 &&
-                      checkoutSuggestion.map((s, i) => (
-                        <span key={i}>{s}</span>
-                      ))}
-                  </div>
-                </div>
-              </div>
-              {bogeyWarning && (
-                <div className="bogey-warning-display">💡 {bogeyWarning}</div>
-              )}
-            </div>
-          )}
-
-          {['scoring_practice_t20', 'checkout_challenge'].includes(
-            gameMode,
-          ) && (
-            <div className="generic-scoreboard">
-              <div className="cricket-header">
-                <h2 className="cricket-title">
-                  {gameMode.replace(/_/g, ' ').toUpperCase()}
-                </h2>
-                <div className="game-controls">
-                  <button className="control-btn" onClick={resetGame}>
-                    RESET
-                  </button>
-                  <button
-                    className="control-btn undo-btn"
-                    onClick={undoLastDart}
-                    disabled={gameHistory.length === 0}
-                  >
-                    UNDO
-                  </button>
-                </div>
-              </div>
-              <div className="practice-scoreboard">
-                <div className="practice-player practice-player-solo">
-                  <div className="player-name">{gameState.player1.name}</div>
-                  {gameMode === 'scoring_practice_t20' ? (
-                    <>
-                      <div className="practice-score">
-                        {gameState.player1.score}
-                      </div>
-                      <div className="target-display">
-                        Darts Remaining: {gameState.player1.dartsRemaining}
-                      </div>
-                      {isGameWon() === 'Practice Over' && (
-                        <div className="practice-summary">
-                          <div className="practice-summary-stat">
-                            T20s Hit: {gameState.player1.t20s}
-                          </div>
-                          <div className="practice-summary-stat">
-                            S20s Hit: {gameState.player1.s20s}
-                          </div>
-                          <div className="practice-summary-stat">
-                            Other: {gameState.player1.otherHits}
-                          </div>
-                        </div>
-                      )}
-                    </>
-                  ) : (
-                    <>
-                      <div className="practice-score">
-                        {gameState.player1.currentTarget}
-                      </div>
-                      <div className="target-display">
-                        Successful Checkouts:{' '}
-                        {gameState.player1.successfulCheckouts}
-                      </div>
-                    </>
-                  )}
-                </div>
-              </div>
-            </div>
-          )}
-
-          {['bobs_27', 'around_the_clock', 'shanghai'].includes(gameMode) && (
-            <div className="generic-scoreboard">
-              <div className="cricket-header">
-                <h2 className="cricket-title">
-                  {gameMode.replace(/_/g, ' ').toUpperCase()}
-                </h2>
-                <div className="game-controls">
-                  <button className="control-btn" onClick={resetGame}>
-                    RESET
-                  </button>
-                  <button className="control-btn" onClick={() => nextPlayer()}>
-                    NEXT PLAYER
-                  </button>
-                  <button
-                    className="control-btn undo-btn"
-                    onClick={undoLastDart}
-                    disabled={gameHistory.length === 0}
-                  >
-                    UNDO
-                  </button>
-                </div>
-              </div>
-              <div className="last-dart-display">
-                {lastDart ? (
-                  <div>
-                    <span className="last-dart-text">LAST DART:</span>
-                    <span className="last-dart-value">{lastDart}</span>
-                  </div>
-                ) : (
-                  <span className="last-dart-text">
-                    Waiting for dart throw...
-                  </span>
-                )}
-              </div>
-              <div className="practice-scoreboard">
-                <div
-                  className={`practice-player ${currentPlayer === 1 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player1.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 1 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 1 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="practice-score">
-                    {gameState.player1.score}
-                  </div>
-                  <div className="target-display">
-                    TARGET:{' '}
-                    {gameMode === 'bobs_27'
-                      ? `D${gameState.player1.currentTarget}`
-                      : gameMode === 'around_the_clock'
-                        ? gameState.player1.currentTarget > 20
-                          ? 'BULL'
-                          : gameState.player1.currentTarget
-                        : gameMode === 'shanghai'
-                          ? gameState.player1.currentTarget
-                          : ''}
-                  </div>
-                </div>
-                <div
-                  className={`practice-player ${currentPlayer === 2 ? 'active' : ''}`}
-                >
-                  <div className="player-name">{gameState.player2.name}</div>
-                  <div
-                    className={`dart-indicator ${currentPlayer === 2 ? 'active' : ''}`}
-                  >
-                    {currentPlayer === 2 ? `DART ${dartCount}/3` : 'WAITING'}
-                  </div>
-                  <div className="practice-score">
-                    {gameState.player2.score}
-                  </div>
-                  <div className="target-display">
-                    TARGET:{' '}
-                    {gameMode === 'bobs_27'
-                      ? `D${gameState.player2.currentTarget}`
-                      : gameMode === 'around_the_clock'
-                        ? gameState.player2.currentTarget > 20
-                          ? 'BULL'
-                          : gameState.player2.currentTarget
-                        : gameMode === 'shanghai'
-                          ? gameState.player2.currentTarget
-                          : ''}
-                  </div>
-                </div>
-              </div>
-            </div>
-          )}
-        </>
+        <ScoreboardUI
+          activeGameSection={activeGameSection}
+          toggleSection={toggleSection}
+          standardGames={standardGames}
+          partyGames={partyGames}
+          practiceModes={practiceModes}
+          gameMode={gameMode}
+          setGameModeAndReset={setGameModeAndReset}
+          isGameWon={isGameWon}
+          KillerScoreboard={KillerScoreboard}
+          gameState={gameState}
+          currentPlayer={currentPlayer}
+          dartCount={dartCount}
+          cricketNumbers={cricketNumbers}
+          getMarkDisplay={getMarkDisplay}
+          resetGame={resetGame}
+          nextPlayer={nextPlayer}
+          undoLastDart={undoLastDart}
+          goBackTurn={goBackTurn}
+          lastDart={lastDart}
+          gameHistory={gameHistory}
+          checkoutSuggestion={checkoutSuggestion}
+          bogeyWarning={bogeyWarning}
+        />
       )}
-      <div className="dartboard-component">
-        <div className="dartboard-glow"></div>
-        <div className="dartboard-svg-container">
-          <svg className="dartboard-svg" viewBox="0 0 500 500">
-            <defs>
-              <radialGradient id="gradient-a" cx="50%" cy="50%" r="50%">
-                <stop offset="0%" style={{ stopColor: '#f4e7a8' }} />
-                <stop offset="100%" style={{ stopColor: '#d4bc3c' }} />
-              </radialGradient>
-              <radialGradient id="gradient-b" cx="50%" cy="50%" r="50%">
-                <stop offset="0%" style={{ stopColor: '#3a3a3a' }} />
-                <stop offset="100%" style={{ stopColor: '#050505' }} />
-              </radialGradient>
-              <radialGradient id="gradient-g" cx="50%" cy="50%" r="50%">
-                <stop offset="0%" style={{ stopColor: '#22c55e' }} />
-                <stop offset="100%" style={{ stopColor: '#15803d' }} />
-              </radialGradient>
-              <radialGradient id="gradient-r" cx="50%" cy="50%" r="50%">
-                <stop offset="0%" style={{ stopColor: '#ef4444' }} />
-                <stop offset="100%" style={{ stopColor: '#b91c1c' }} />
-              </radialGradient>
-              <filter id="bevel" x="-50%" y="-50%" width="200%" height="200%">
-                <feGaussianBlur
-                  in="SourceAlpha"
-                  stdDeviation="1"
-                  result="blur"
-                />
-                <feOffset in="blur" dx="2" dy="2" result="offsetBlur" />
-                <feSpecularLighting
-                  in="blur"
-                  surfaceScale="3"
-                  specularConstant=".75"
-                  specularExponent="10"
-                  lightingColor="#ffffff"
-                  result="specOut"
-                >
-                  <fePointLight x="-5000" y="-10000" z="20000" />
-                </feSpecularLighting>
-                <feComposite
-                  in="specOut"
-                  in2="SourceAlpha"
-                  operator="in"
-                  result="specOut"
-                />
-                <feComposite
-                  in="SourceGraphic"
-                  in2="specOut"
-                  operator="arithmetic"
-                  k1="0"
-                  k2="1"
-                  k3="1"
-                  k4="0"
-                  result="litPaint"
-                />
-                <feMerge>
-                  <feMergeNode in="offsetBlur" />
-                  <feMergeNode in="litPaint" />
-                </feMerge>
-              </filter>
-              <filter
-                id="bevel-hover"
-                x="-50%"
-                y="-50%"
-                width="200%"
-                height="200%"
-              >
-                <feGaussianBlur
-                  in="SourceAlpha"
-                  stdDeviation="2"
-                  result="blur"
-                />
-                <feOffset in="blur" dx="3" dy="3" result="offsetBlur" />
-                <feSpecularLighting
-                  in="blur"
-                  surfaceScale="5"
-                  specularConstant="1"
-                  specularExponent="20"
-                  lightingColor="#ffffff"
-                  result="specOut"
-                >
-                  <fePointLight x="-5000" y="-10000" z="20000" />
-                </feSpecularLighting>
-                <feComposite
-                  in="specOut"
-                  in2="SourceAlpha"
-                  operator="in"
-                  result="specOut"
-                />
-                <feComposite
-                  in="SourceGraphic"
-                  in2="specOut"
-                  operator="arithmetic"
-                  k1="0"
-                  k2="1"
-                  k3="1"
-                  k4="0"
-                  result="litPaint"
-                />
-                <feMerge>
-                  <feMergeNode in="offsetBlur" />
-                  <feMergeNode in="litPaint" />
-                </feMerge>
-              </filter>
-              <filter
-                id="board-shadow"
-                x="-50%"
-                y="-50%"
-                width="200%"
-                height="200%"
-              >
-                <feDropShadow
-                  dx="10"
-                  dy="10"
-                  stdDeviation="10"
-                  floodColor="#000000"
-                  floodOpacity="0.5"
-                />
-              </filter>
-            </defs>
-            <g transform="translate(50, 50)">
-              {generateSegments()}
-              <BullseyeGlowWrapper>
-                <circle
-                  cx={center}
-                  cy={center}
-                  r={radii.bullOuter}
-                  className="segment color-g"
-                  data-score="single-bull"
-                  onClick={() => handleDartThrow('single-bull')}
-                />
-              </BullseyeGlowWrapper>
-              <BullseyeGlowWrapper>
-                <circle
-                  cx={center}
-                  cy={center}
-                  r={radii.bullInner}
-                  className="segment color-r"
-                  data-score="double-bull"
-                  onClick={() => handleDartThrow('double-bull')}
-                />
-              </BullseyeGlowWrapper>
-            </g>
-            <g transform="translate(50, 50)">{generateNumbers()}</g>
-          </svg>
-        </div>
-      </div>
+      <DartboardSVG
+        center={center}
+        radii={radii}
+        generateSegments={generateSegments}
+        generateNumbers={generateNumbers}
+        handleDartThrow={handleDartThrow}
+        BullseyeGlowWrapper={BullseyeGlowWrapper}
+      />
     </div>
   )
 }
